Did a bunch of refactoring
