+++ JamfSync/Model/JsonObjects.swift
@@ -239,41 +239,3 @@ struct JsonInitiateUpload: Decodable {
 struct JsonJamfProVersion: Decodable {
     let version: String
 }
-
-// Added - BIG-RAT
-struct Headers {
-    static var canonical = ""
-    static var signed    = ""
-}
-
-class UploadTime {
-    var start: Int
-    var end: Int
-    
-    init(start: Int, end: Int) {
-        self.start = start
-        self.end = end
-    }
-    
-    func total() -> String {
-        let totalSeconds = end - start
-        var totalTime = ""
-        if totalSeconds >= 3600 {
-            let hours = totalSeconds / 3600
-            let unit = (hours == 1 ) ? "hour ":"hours "
-            totalTime = "\(hours) \(unit)"
-        }
-        if totalSeconds >= 60 {
-            let minutes = (totalSeconds % 3600) / 60
-            let unit = (minutes == 1 ) ? "minute ":"minutes "
-            totalTime = totalTime + "\(minutes) \(unit)"
-        }
-        if totalSeconds >= 0 {
-            let seconds = (totalSeconds % 3600) % 60
-            let unit = (seconds == 1 ) ? "second":"seconds"
-            totalTime = totalTime + "\(seconds) \(unit)"
-        }
-        return totalTime
-    }
-}
-var uploadTime: UploadTime = UploadTime(start: 0, end: 0)
